fix: opt code
