+++ packages/core/src/shaderlib/pbr/pbr_helper.glsl
@@ -177,7 +177,7 @@ void initMaterial(out Material material, inout Geometry geometry){
             specularAO = saturate( pow( geometry.dotNV + diffuseAO, exp2( - 16.0 * material.roughness - 1.0 ) ) - 1.0 + diffuseAO );
         #endif
 
-        #ifdef ENABLE_SSAO                    
+        #ifdef SCENE_ENABLE_SSAO                    
             vec4 samplingPositionNDC = camera_ProjMat * camera_ViewMat * vec4( geometry.position, 1.0 );
             vec2 ssaoUV = (samplingPositionNDC.xy / samplingPositionNDC.w) * 0.5 + 0.5;
             float ssao = texture2D(camera_SSAOTexture, ssaoUV).r;
